+++ astroquery/xmatch/tests/test_xmatch.py
@@ -104,9 +104,3 @@ def test_xmatch_query_cat1_table_local(monkeypatch):
         'errHalfMaj', 'errHalfMin', 'errPosAng', 'Jmag', 'Hmag', 'Kmag',
         'e_Jmag', 'e_Hmag', 'e_Kmag', 'Qfl', 'Rfl', 'X', 'MeasureJD']
     assert len(table) == 11
-
-
-@pytest.mark.parametrize('datafile', DATA_FILES.values())
-def test_parse_text(datafile):
-    xm = XMatch()
-    xm._parse_text(datafile)